+++ Meet-My-Lecture/src/main/java/com/springboot/meetMyLecturer/repository/MeetingRequestRepository.java
@@ -4,6 +4,9 @@ import com.springboot.meetMyLecturer.entity.MeetingRequest;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface MeetingRequestRepository extends JpaRepository<MeetingRequest, Long> {
+public interface MeetingRequestRepository extends JpaRepository<MeetingRequest, Integer> {
+
 }
