+++ Meet-My-Lecture/src/main/java/com/springboot/meetMyLecturer/service/UserService.java
@@ -11,6 +11,6 @@ public interface UserService {
 
     List<UserDTO> searchLecturers(String name);
 
-    List<UserDTO> getUserByEmptySlotId(Long slotId);
+    List<UserDTO> getUserByEmptySlotId(int slotId);
 
 }
